{{ cookiecutter.project_slug }}/django_proj/requirements.txt
asgiref==3.5.2
# backports.zoneinfo==0.2.1
Django==4.1.3
psycopg2==2.9.5
sqlparse==0.4.3

django-cors-headers==3.14.0
{% if cookiecutter.configure_celery %}
celery==5.3.0
django-celery-results==2.5.1
django-celery-beat==2.5.0
{% endif %}
django-cleanup==6.0.0
djangorestframework==3.14.0
django-allauth~=0.51.0
dj-rest-auth~=2.2.2

# production dependencies
gunicorn==20.1.0
supervisor==4.2.4